Update camera code and integrate with the repository structure
