+++ artifacts/cdh_em_board/code.py
@@ -1,9 +1,10 @@
 import asyncio
+from tasks.camera_tasks import capture_single_image
 
 async def gathered_task():
     """Runs all tasks in parallel."""
     await asyncio.gather(
-
+        capture_single_image(return_bursts = False)
     )
