+++ artifacts/cdh_em_board/include.json
@@ -1,0 +1,8 @@
+{
+    "src": [
+        "drivers/camera.py",
+        "lib/asyncio",
+        "tasks/camera_tasks.py"
+    ],
+    "unit_tests": []
+}
